Current

+++ basepkg.sh
@@ -8,7 +8,7 @@ machine_arch="$(uname -p)"
 opsys="$(uname)"
 osversion="$(uname -r)"
 pkgtoolversion="$(pkg_add -V)"
-prog="${0##*/}"
+osrelease="$(sh ${SRC}/sys/conf/osrelease.sh)"
 rcsid='$NetBSD: make_basepkg.sh,v 0.01 2016/10/19 15:36:22 uki Exp $'
 utcdate="$(env TZ=UTC LOCALE=C date '+%Y-%m-%d %H:%M')"
 user="${USER:-root}"
@@ -18,8 +18,10 @@ lists="${database}/lists"
 comments="${database}/comments"
 descrs="${database}/descrs"
 deps="${database}/deps"
+tmp_deps="/tmp/culldeps"
 category="base comp etc games man misc text"
 progname=${0##*/}
+target="${category}"
 
 # "extract" option using following function.
 extract_base_binaries() {
@@ -109,16 +111,53 @@ PKGTOOLS_VERSION=${pkgtoolversion}
 _BUILD_INFO_
 }
 
+culc_deps() {
+	grep -E "^$1" ${deps} > /dev/null 2>&1
+	if [ $? -eq 1 ]; then
+		echo "$1:Unknown package dependency." 1>&2
+		return 1
+	fi
+	TMP=`mktemp -q`
+	if [ $? -ne 0 ]; then
+		echo "$0: Can't create temp file, exiting..." 1>&2
+		exit 1
+	fi
+	grep -E "^$1" ${deps} | awk '{print $2}' > ${TMP}
+	# XXX: too many temp files in /tmp
+	cat ${TMP} | while read depend
+	do
+		if [ ! "${depend}" ]; then
+			rm -f ${TMP}
+			return 1
+		fi
+		echo "@pkgdep ${depend}>=${osrelease}" >> ${tmp_deps}
+		if [ "${depend}" = "base-sys-root" ]; then
+			rm ${TMP}
+			return 0
+		fi
+		culc_deps ${depend}
+	done
+}
+
 make_CONTENTS() {
-	if [ -f ./$1/tmp.list ]; then
-		rm -f ./$1/tmp.list
+	TMPFILE=`mktemp -q`
+	if [ $? -ne 0 ]; then
+		echo "$0: Can't create temp file, exiting..."
+		exit 1
 	fi
 	setname=`echo $1 | awk 'BEGIN{FS="/"} {print $1}' | sed 's/\./-/g'`
 	pkgname=`echo $1 | awk 'BEGIN{FS="/"} {print $2}' | sed 's/\./-/g'`
-	echo "@name ${pkgname}-`sh ${SRC}/sys/conf/osrelease.sh`" > ./$1/+CONTENTS
+	echo "@name ${pkgname}-${osrelease}" > ./$1/+CONTENTS
 	echo "@comment Packaged at ${utcdate} UTC by ${user}@${host}" \
 	>> ./$1/+CONTENTS
-	echo "@comment Packaged using ${prog} ${rcsid}" >> ./$1/+CONTENTS
+	echo "@comment Packaged using ${progname} ${rcsid}" >> ./$1/+CONTENTS
+	if [ -f ${tmp_deps} ]; then
+		rm -f ${tmp_deps}
+	fi
+	culc_deps ${pkgname}
+	if [ -f ${tmp_deps} ]; then
+		sort ${tmp_deps} | uniq >> ./$1/+CONTENTS
+	fi
 	echo "@cwd /" >> ./$1/+CONTENTS
 	cat ./$1/${pkgname}.FILES | while read i
 	do
@@ -127,18 +166,15 @@ make_CONTENTS() {
 			filename=`echo ${i} | sed 's%\/%\\\/%g'`
 			awk '$1 ~ /^\.\/'"${filename}"'$/{print $0}' ./work/${setname}/etc/mtree/set.${setname} | \
 			sed 's%^\.\/%%' | \
-			awk '{print "@exec install -d -o root -g wheel -m "substr($5, 6) " "$1}' >> ./$1/tmp.list
+			awk '{print "@exec install -d -o root -g wheel -m "substr($5, 6) " "$1}' >> ${TMPFILE}
 		elif [ ${filetype} = cannot ]; then
 			continue
 		else
-			echo ${i} >> ./$1/tmp.list
+			echo ${i} >> ${TMPFILE}
 		fi
 	done
-	if [ ! -f ./$1/tmp.list ]; then
-		return 1
-	fi
-	sort ./$1/tmp.list >> ./$1/+CONTENTS
-	rm -f ./$1/tmp.list
+	sort ${TMPFILE} >> ./$1/+CONTENTS
+	rm -f ${TMPFILE}
 }
 
 make_DESC_and_COMMENT() {
@@ -162,11 +198,11 @@ make_PKG() {
 	  mkdir -p ${PACKAGES}/${setname}
 	fi
 	mv ./${pkgname}.tgz \
-	${PACKAGES}/${setname}/${pkgname}-`sh ${SRC}/sys/conf/osrelease.sh`.tgz
+	${PACKAGES}/${setname}/${pkgname}-${osrelease}.tgz
 }
 
 make_packages() {
-	for i in ${category}
+	for i in ${target}
 	do
 		for j in `ls ./${i} | grep -E '^[a-z]+'`
 		do
@@ -183,10 +219,16 @@ make_packages() {
 clean_packages() {
 	for i in ${category}
 	do
+		if [ ! -d ${PACKAGES}/${i} ]; then
+			continue
+		fi
 		ls ${PACKAGES}/${i} | grep -E 'tgz$' | \
 		xargs -I % rm -f ${PACKAGES}/${i}/%
 		rmdir ${PACKAGES}/${i}
 	done
+	if [ ! -d ${PACKAGES} ]; then
+		return 1
+	fi
 	rmdir ${PACKAGES}
 }
 
@@ -201,7 +243,6 @@ clean_categories() {
 			test -f ./${i}/${j}/+COMMENT && rm -f ./${i}/${j}/+COMMENT
 			test -f ./${i}/${j}/+CONTENTS && rm -f ./${i}/${j}/+CONTENTS
 			test -f ./${i}/${j}/+DESC && rm -f ./${i}/${j}/+DESC
-			test -f ./${i}/${j}/tmp.list && rm -f ./${i}/${j}/tmp.list
 			test -f ./${i}/${j}/${j}.FILES && rm -f ./${i}/${j}/${j}.FILES
 			rmdir ./${i}/${j}
 		done
@@ -223,8 +264,9 @@ Usage: ${progname} [--sets sets] [--src src] [--pkgsrc pkgsrc]
     list                Create packages list.
     pkg                 Create packages.
     all                 Running dir,list,pkg options.
-	cleanpkg            Remove all packages.
-    cleanall            Remove all packages and created directories.
+    cleanpkg            Remove all packages.
+    cleandir            Remove all categorized directories.
+    cleanall            Remove all packages and categorized directories.
 
  Options:
     --help              Show this message and exit.
@@ -234,6 +276,8 @@ Usage: ${progname} [--sets sets] [--src src] [--pkgsrc pkgsrc]
                         [Default: /usr/src]
     --pkg               Set packages root directory; sets a PACKAGES pattern.
                         [Default: ./packages]
+    --target            Set target packaging category.
+                        [Default: "${category}"]
 
 _usage_
 	exit 1
@@ -270,6 +314,15 @@ do
 				exit 1
 			fi
 			PACKAGES=$2
+			shift
+			shift
+			;;
+		'--target' )
+			if [ -z $2 ]; then
+				echo "What is $1 parameter?" 1>&2
+				exit 1
+			fi
+			target="$2"
 			shift
 			shift
 			;;
